add more api, optimize extractor, prepare new training data

+++ parser/crawler.py
@@ -13,19 +13,25 @@ class PageCrawler(object):
         self.logger = get_logger(self.__class__.__name__)
 
     def process(self, urls):
+        urls = list(set(urls))
         result = {}
-        # use multi thread to crawl pages
-        pool = Pool(cpu_count() * 2)
-        pool_results = pool.map(self._crawl_page, urls)
-        # get results
-        for r in pool_results:
-            result.update(r)
+        if len(urls) > 2:
+            # use multi thread to crawl pages
+            pool = Pool(cpu_count() * 2)
+            pool_results = pool.map(self._crawl_page, urls)
+            # get results
+            for r in pool_results:
+                result.update(r)
 
-        pool.terminate()
+            pool.terminate()
+        else:
+            for url in urls:
+                result.update(self._crawl_page(url))
+
         return result
 
     def _crawl_page(self, url):
-        self.logger.debug('Start crawl %s...' + url)
+        self.logger.debug('Start crawl %s...' % url)
         result = {
             url: {
                 'content': '',
@@ -51,7 +57,7 @@ class PageCrawler(object):
             result[url]['error'] = True
             result[url]['message'] = 'url is empty'
 
-        self.logger.debug('End crawl %s...' + url)
+        self.logger.debug('End crawl %s...' % url)
         return result
 
 
@@ -63,31 +69,41 @@ class PageCrawlerWithStorage(object):
 
     def process(self, urls):
         result = {}
+        urls = list(set(urls))
         # get crawled pages
         for page in self.storage.find({'_id': {'$in': urls}}):
             if page.get('crawled_date'):
                 self.logger.debug('Page was crawled: ' + page['_id'])
                 result[page['_id']] = page
+
+        self.logger.info("Num of crawled urls: %s" % len(result))
         # filter crawled page
         urls = [u for u in urls if u not in result]
+
+        self.logger.info("Remain haven't crawled urls: %s" % len(urls))
 
         if not urls:
             self.logger.info('All urls has been crawled')
             return result
 
-        # use multi thread to crawl pages
-        pool = Pool(cpu_count() * 2)
-        self.logger.debug('Have to crawl these urls: %s' % urls)
-        pool_results = pool.map(self._crawl_page, urls)
-        # get results
-        for r in pool_results:
-            result.update(r)
+        if len(urls) > 2:
+            # use multi thread to crawl pages
+            pool = Pool(cpu_count() * 2)
+            self.logger.debug('Have to crawl these urls: %s' % urls)
+            pool_results = pool.map(self._crawl_page, urls)
+            # get results
+            for r in pool_results:
+                result.update(r)
 
-        pool.terminate()
+            pool.terminate()
+        else:
+            for url in urls:
+                result.update(self._crawl_page(url))
+
         return result
 
     def _crawl_page(self, url):
-        self.logger.debug('Start crawl %s...' + url)
+        self.logger.debug('Start crawl %s...' % url)
         result = {
             'content': '',
             'error': False,
@@ -96,7 +112,7 @@ class PageCrawlerWithStorage(object):
         if url:
             # check database
             page = self.storage.find_one({'_id': url})
-            if page.get('crawled_date'):
+            if page and page.get('crawled_date'):
                 self.logger.debug('Page was crawled (2nd check): ' + page['_id'])
                 return {url: self.storage.find_one({'_id': url})}
 
@@ -123,5 +139,5 @@ class PageCrawlerWithStorage(object):
         result['content'] = get_unicode(result['content'])
         self.logger.info('Update crawled page to db...')
         self.storage.update_one({'_id': url}, {'$set': result}, True)
-        self.logger.debug('End crawl %s...' + url)
+        self.logger.debug('End crawl %s...' % url)
         return {url: self.storage.find_one({'_id': url})}
